% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data}
\format{
A data frame with 144 rows and 6 variables
\describe{
 \item{RIVER}{River fish was found in}
 \item{MILE}{distance down river fish was found}
 \item{SPECIES}
 \item{LENGTH}{length in cm}
 \item{WEIGHT}{weight in grams}
 \item{DDT}{DDT levels in the fish}
 }
}
\usage{
ddt
}
\description{
A data set with information on DDT levels in different species of fish in the Tennesse River and its tributaries
}
\keyword{datasets}
