% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin_lab5.R
\name{mybin}
\alias{mybin}
\title{Lab 5 Function - Binomial}
\usage{
mybin(iter = 100, n = 10, p = 0.5, tab = FALSE)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size in each iteration}

\item{p}{probablity of success}

\item{tab}{gives a table of successes if set to TRUE}
}
\value{
Returns a barplot of a series of binomial samples. Can also return a table.
}
\description{
Lab 5 Function - Binomial
}
\examples{
\dontrun{mybin(iter=50, n=10, p=0.75, tab=TRUE)}
}
