% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2_lab9.R
\name{myboot2}
\alias{myboot2}
\title{Lab 9 function: myboot2()}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{data being sampled}

\item{fun}{a function that will give a statistic for the data}

\item{alpha}{alpha for the confidence interval}

\item{cx}{font size of the numbers on the histogram}

\item{...}{for adding more customization to the histogram}
}
\value{
Returns a histogram of the statistic in "fun". Also can return the values for the confidence interval, the values for all the data used, and the function that was used to find the statistic.
}
\description{
Lab 9 function: myboot2()
}
\examples{
\dontrun{myboot2(iter = 10000, x, fun = "mean", alpha = 0.05)}
\dontrun{myboot2(iter = 5000, x, fun = "var", alpha = 0.20)}

}
