% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci_lab11.R
\name{myci}
\alias{myci}
\title{Lab 11 Function - Confidence Interval for the Mean}
\usage{
myci(x)
}
\arguments{
\item{x}{the sample data}
}
\value{
returns a 95 percent confidence interval for the population mean
}
\description{
Lab 11 Function - Confidence Interval for the Mean
}
