% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve_lab6.R
\name{myncurve}
\alias{myncurve}
\title{Lab 6 Function - Normal curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the normal curve}

\item{sigma}{the sd of the normal curve}

\item{a}{an x value within the curve}
}
\value{
returns a plot of the normal curve that is shaded from effectively -infinity to a. Also returns the area of the shaded area to the command line.
}
\description{
Lab 6 Function - Normal curve
}
\examples{
\dontrun{myncurve(mu=0, sigma=3, a=1)}
}
