% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction_lab4.R
\name{myquadplot}
\alias{myquadplot}
\title{Lab 4 function}
\usage{
myquadplot(x)
}
\arguments{
\item{x}{is an independent value}
}
\value{
y hat value
}
\description{
Lab 4 function
}
\examples{
\dontrun{myquadplot(x = 11)}
}
